requirements.txt
config==0.3.9
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==2.2.6
django-bootstrap-modal-forms==1.4.4
django-bootstrap3==10.0.1
django-heroku==0.3.1
django-registration==2.4.1
django-tinymce==2.8.0
djangorestframework==3.10.3
gunicorn==19.9.0
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.8.4
pyperclip==1.7.0
python-decouple==3.1
pytz==2018.5
sqlparse==0.3.0
whitenoise==4.0
